% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{cmn.lst}
\alias{cmn.lst}
\title{cmn.lst}
\format{
An object of class \code{character} of length 7904.
}
\usage{
cmn.lst
}
\description{
Lista completa unica dei Comuni Italiani
}
\keyword{datasets}
