% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estrazioni.R
\name{estrai_gridpop}
\alias{estrai_gridpop}
\title{estrai_gridpop}
\usage{
estrai_gridpop(x, s, out = "m", ...)
}
\arguments{
\item{x}{Vettore di codici ISTAT Comune}

\item{s}{Segmento della popolazione (inserire lettera iniziale dei nomi in elenco \code{fb_pop.lst})}

\item{out}{Carattere che indica che tipo di oggetto ritornare; 
uno fra: 'd' una tabella data.table, 's' oggetto spaziale sf, 'm' mappa leaflet}

\item{...}{ulteriori argomenti da passare alla funzione \code{basemap} del pacchetto \code{masteRfun}}
}
\value{
Una tabella "data.table", un oggetto spaziale "sf", oppure una mappa interattiva "leaflet"
}
\description{
Estrae i valori di griglia per uno o piu Comuni ed uno specifico segmento di popolazione
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
