% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati.R
\docType{data}
\name{fb_centroidi}
\alias{fb_centroidi}
\title{fb_centroidi}
\format{
Una data.table con i campi seguenti:
\describe{
  \item{\code{ref}}{ segmento di popolazione }
  \item{\code{CMN}}{ Codice ISTAT del Comune }
  \item{\code{x_lon}}{ Longitudine del centroide pesato }
  \item{\code{y_lat}}{ Latitudine del centroide pesato }
}
}
\usage{
fb_centroidi
}
\description{
Centroidi pesati rispetto alla popolazione totale o ad un suo particolare segmento (vedi elenco \code{fb_pop.lst})
}
\keyword{datasets}
