% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{fb_pop.lst}
\alias{fb_pop.lst}
\title{Lista dei segmenti di popolazione presenti nei microdati Facebook}
\format{
An object of class \code{character} of length 7.
}
\usage{
fb_pop.lst
}
\description{
Lista dei segmenti di popolazione presenti nei microdati Facebook
}
\keyword{datasets}
