% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masteRgridpop.R
\docType{package}
\name{masteRgridpop}
\alias{masteRgridpop}
\title{masteRgridpop}
\description{
Pacchetto $R$ contenente funzioni per richiedere valori di popolazione, totale e per specifici segmenti di interesse, a livello di griglie microterritoriali.
}
\details{
Le sorgenti dei dati sono: Facebook "Data For Good", EU Progetto "Corinne"
}
