% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{tipi_centroidi}
\alias{tipi_centroidi}
\title{tipi_centroidi}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 6 rows and 5 columns.
}
\usage{
tipi_centroidi
}
\description{
Elenco delle varie tipologie di centroidi presenti nella tabella comuni
}
\keyword{datasets}
