% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati.R
\docType{data}
\name{yb}
\alias{yb}
\title{yb}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 7904 rows and 2 columns.
}
\usage{
yb
}
\description{
Confini Comunali ridotti al 10%
}
\keyword{datasets}
