% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati.R
\docType{data}
\name{yc}
\alias{yc}
\title{yc}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 7904 rows and 18 columns.
}
\usage{
yc
}
\description{
Tabella Comuni ridotta
}
\keyword{datasets}
